adding react-router-dom, adding a new Canvas component to house data needed for SoundList and for a future component that plays a specific track
