+++ src/App.js
@@ -1,20 +1,32 @@
 import React from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import "./App.scss";
 import Footer from "./components/Footer";
 import Header from "./components/Header";
-import SoundList from "./components/SoundList";
+import Canvas from "./components/Canvas";
 
 class App extends React.Component {
     render() {
         return (
-            <div className="App">
-                <Header />
-                <div className="container">
-                    <SoundList />
+            <Router>
+                <div className="App">
+                    <Header />
+                    <div className="container">
+                        <Switch>
+                        
+                            <Route exact path="/">
+                                <Canvas />
+                            </Route>
+                            
+                            <Route path="/test">
+                                hello
+                            </Route>
+                        
+                        </Switch>
+                    </div>
+                    <Footer />
                 </div>
-
-                <Footer />
-            </div>
+            </Router>
         );
     }
 }
